+++ Sttp.Core/Codec/CommandObjects/Metadata Request/CommandBeginMetadataResponse.cs
@@ -10,7 +10,7 @@ namespace Sttp.Codec
         : DocumentObject<CommandBeginMetadataResponse>
     {
         [DocumentField()]
-        public int BinaryChannelCode { get; private set; }
+        public ulong BinaryChannelCode { get; private set; }
         [DocumentField()]
         public Guid EncodingMethod { get; private set; }
         [DocumentField()]
@@ -22,7 +22,7 @@ namespace Sttp.Codec
         [DocumentField()]
         public List<MetadataColumn> Columns { get; private set; }
 
-        public CommandBeginMetadataResponse(int binaryChannelCode, Guid encodingMethod, Guid runtimeID, long versionNumber, string tableName, List<MetadataColumn> columns)
+        public CommandBeginMetadataResponse(ulong binaryChannelCode, Guid encodingMethod, Guid runtimeID, long versionNumber, string tableName, List<MetadataColumn> columns)
         {
             BinaryChannelCode = binaryChannelCode;
             EncodingMethod = encodingMethod;
